requirements.txt
beautifulsoup4==4.14.2
deepeval==3.7.2
groq==0.34.1
openai==2.8.1
pandas==2.3.3
Pillow>=10.0.0
protobuf==6.33.1
pymysql==1.1.2
PyPDF2==3.0.1
python-dotenv==1.2.1
python_docx==1.1.2
Requests==2.32.5
SQLAlchemy==2.0.44
streamlit==1.45.1
streamlit_feedback==0.1.4
tenacity==8.5.0


# --------- core data & cloud ---------
#pandas>=2.0.0
#numpy>=1.24.0
#PyMySQL>=1.1.0
#mysql-connector-python>=8.3.0
#google-cloud-bigquery>=3.12.0
#SQLAlchemy>=2.0.0

# --------- AI / LLM ---------
#openai>=1.10.0
#groq>=0.9.0
#transformers>=4.38.0
#deepeval>=0.21.0
#langchain
##langchain-community

# --------- document handling ---------
#PyPDF2>=3.0.0
#python-docx>=1.1.0#
#Pillow>=10.0.0

# --------- web app ---------
#streamlit>=1.32.0
#streamlit-feedback>=0.1.3

# --------- environment & utilities ---------
#python-dotenv>=1.0.0
openpyxl>=3.1.0
